// src/enums/EJwtToken.ts
export enum EJwtToken {
  // ACCESS_TOKEN_EXPIRATION = 3600,
  ACCESS_TOKEN_EXPIRATION = 30, // 30s for test
  REFRESH_TOKEN_EXPIRATION = 2592000,
}

export enum EJwtExpirationErrorCode {
  ACCESS_TOKEN_EXPIRED = 1001,
  REFRESH_TOKEN_EXPIRED = 1002,
  NO_TOKEN_PROVIDED = 1003,
  TOKEN_REUSE_DETECTED = 1004,
  GRANT_REVOKED = 1005,
}